Rename JsDashboard -> DashboardTrigger
