+++ app/src/main/java/com/jaspersoft/android/jaspermobile/webview/dashboard/bridge/DashboardTrigger.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2014 TIBCO Software, Inc. All rights reserved.
+ * Copyright © 2015 TIBCO Software, Inc. All rights reserved.
  * http://community.jaspersoft.com/project/jaspermobile-android
  *
  * Unless you have purchased a commercial license agreement from Jaspersoft,
@@ -24,42 +24,23 @@
 
 package com.jaspersoft.android.jaspermobile.webview.dashboard.bridge;
 
-import android.webkit.WebView;
-
 /**
- * Introduces hardcoded Javascript calls.
+ * Interfaces which represent communication native code with dashboard abstraction on Javascript side.
  *
  * @author Tom Koptel
- * @since 2.0
+ * @since 2.1
  */
-public final class JsDashboardImpl implements JsDashboard {
-    private final WebView webView;
-
-    private JsDashboardImpl(WebView webView) {
-        this.webView = webView;
-    }
-
-    public static JsDashboardImpl with(WebView webView) {
-        if (webView == null) {
-            throw new IllegalArgumentException("WebView reference should not be null");
-        }
-        return new JsDashboardImpl(webView);
-    }
-
-    public void refreshDashlet() {
-        webView.loadUrl(assembleUri("MobileDashboard.refresh()"));
-    }
-
-    public void minimizeDashlet() {
-        webView.loadUrl(assembleUri("MobileDashboard.minimizeDashlet()"));
-    }
-
-    @Override
-    public void refreshDashboard() {
-        webView.loadUrl(assembleUri("MobileDashboard.refresh()"));
-    }
-
-    private String assembleUri(String command) {
-        return "javascript:" + command;
-    }
+public interface DashboardTrigger {
+    /**
+     * Refreshes maximized dashlet.
+     */
+    void refreshDashlet();
+    /**
+     * Refreshes all dashlets.
+     */
+    void refreshDashboard();
+    /**
+     * Minimize latest maximized dashlet.
+     */
+    void minimizeDashlet();
 }
